% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfFlowList-methods.R,
%   R/ncdfFlowSet-accessors.R
\name{filter,ncdfFlowList,filter-method}
\alias{filter,ncdfFlowList,filter-method}
\alias{filter,ncdfFlowList,filterList-method}
\alias{filter,ncdfFlowList,list-method}
\alias{length,ncdfFlowList-method}
\alias{sampleNames,ncdfFlowList-method}
\alias{phenoData,ncdfFlowList-method}
\alias{phenoData<-,ncdfFlowList,AnnotatedDataFrame-method}
\alias{pData,ncdfFlowList-method}
\alias{pData<-,ncdfFlowList,data.frame-method}
\alias{colnames,ncdfFlowList-method}
\alias{colnames<-,ncdfFlowList-method}
\alias{markernames,ncdfFlowList-method}
\alias{markernames<-,ncdfFlowList-method}
\alias{compensate,ncdfFlowSet,ANY-method}
\alias{compensate,flowSet,data.frame-method}
\alias{compensate,ncdfFlowSet,list-method}
\alias{transform,ncdfFlowSet-method}
\alias{sampleNames<-}
\alias{sampleNames<-,ncdfFlowSet,ANY-method}
\alias{colnames<-}
\alias{colnames<-,ncdfFlowSet,ANY-method}
\alias{colnames<-,ncdfFlowSet-method}
\alias{keyword,ncdfFlowSet,list-method}
\alias{keyword<-,ncdfFlowSet,list-method}
\title{Accessors compatible with those for \code{flowSet}}
\usage{
\S4method{filter}{ncdfFlowList,filter}(
  x,
  filter,
  method = "missing",
  sides = "missing",
  circular = "missing",
  init = "missing"
)

\S4method{filter}{ncdfFlowList,filterList}(
  x,
  filter,
  method = "missing",
  sides = "missing",
  circular = "missing",
  init = "missing"
)

\S4method{filter}{ncdfFlowList,list}(
  x,
  filter,
  method = "missing",
  sides = "missing",
  circular = "missing",
  init = "missing"
)

\S4method{length}{ncdfFlowList}(x)

\S4method{sampleNames}{ncdfFlowList}(object)

\S4method{phenoData}{ncdfFlowList}(object)

\S4method{pData}{ncdfFlowList}(object)

\S4method{pData}{ncdfFlowList,data.frame}(object) <- value

\S4method{colnames}{ncdfFlowList}(x)

\S4method{colnames}{ncdfFlowList}(x) <- value

\S4method{markernames}{ncdfFlowList}(object)

\S4method{markernames}{ncdfFlowList}(object) <- value

\S4method{compensate}{ncdfFlowSet,ANY}(x, spillover)

\S4method{compensate}{flowSet,data.frame}(x, spillover)

\S4method{compensate}{ncdfFlowSet,list}(x, spillover)

\S4method{transform}{ncdfFlowSet}(x, translist, ...)

\S4method{sampleNames}{ncdfFlowSet,ANY}(object) <- value

\S4method{colnames}{ncdfFlowSet}(x) <- value

\S4method{keyword}{ncdfFlowSet,list}(object, keyword)

\S4method{keyword}{ncdfFlowSet,list}(object) <- value
}
\arguments{
\item{x}{\code{ncdfFlowSet}}

\item{filter}{\code{filter} to be applied}

\item{method}{\code{missing} not used}

\item{sides}{\code{missing} not used}

\item{circular}{\code{missing} not used}

\item{init}{\code{missing} not used}

\item{object}{\code{ncdfFlowList}}

\item{value}{\code{character} vector}

\item{spillover}{spillover matrix}

\item{translist}{a 'transformList' object or a list of 'transformList' objects}

\item{...}{other arguments}

\item{keyword}{list}
}
\description{
Accessors compatible with those for \code{flowSet}
}
