% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getFileName}
\alias{getFileName}
\title{get the cdf file name associated with ncdfFlowSet object}
\usage{
getFileName(ncfs)
}
\arguments{
\item{ncfs}{\code{ncdfFlowSet}}
}
\value{
\code{character}
}
\description{
get the cdf file name associated with ncdfFlowSet object
}

