% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getIndices,ncdfFlowSet,character-method}
\alias{getIndices}
\alias{getIndices,ncdfFlowSet,character-method}
\title{extract the event indices of one or multiple samples from ncdfFlowSet}
\usage{
\S4method{getIndices}{ncdfFlowSet,character}(obj, y)
}
\arguments{
\item{obj}{\code{ncdfFlowSet} object}

\item{y}{\code{character} sample name}
}
\value{
a logical vector.
}
\description{
For internal use.
}
\examples{
data(GvHD)
nc <- ncdfFlowSet(GvHD[1:2])
sn <- sampleNames(nc)[1]
nrow(nc[[sn]])
getIndices(nc, sn) #initial index is NA
#subset with filter
morphGate <- norm2Filter("FSC-H", "SSC-H", filterId = "MorphologyGate",scale = 2)
nc1 <- Subset(nc, morphGate)
ind <- getIndices(nc1, sn)
all.equal(sum(ind), nrow(nc1[[sn]]))
initIndices(nc1)
getIndices(nc1, sn) #reset indices
}

