% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{initIndices,ncdfFlowSet-method}
\alias{initIndices,ncdfFlowSet-method}
\title{initialize the event indices for the entire ncdfFlowSet with NA}
\usage{
\S4method{initIndices}{ncdfFlowSet}(x, y)
}
\description{
initialize the event indices for the entire ncdfFlowSet with NA
}

