% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{package}
\name{ncdfFlow}
\alias{ncdfFlow}
\title{ncdfFlow: A package that provides CDF storage based flow cytometry data analysis.}
\description{
ncdfFlow: A package that provides CDF storage based flow cytometry data analysis.
}
\details{
Define important flow cytometry data classes:
    \code{\link[ncdfFlow:ncdfFlowSet-class]{ncdfFlowSet}}( a subclass of \code{\link[flowCore:flowSet-class]{flowSet}})
and \code{\link[ncdfFlow:ncdfFlowList-class]{ncdfFlowList}}(a list of ncdfFlowSet object) and their accessors.

Provide important compensation,transformation,filter,gating,subsetting,splitting functions for data analysis of large volumns of flow cytometry data that is too big to be held in memory.


\tabular{ll}{
  Package: \tab ncdfFlow\cr
  Version: \tab 2.9.24\cr
  Date:\tab 2014-04-16\cr
  Depends: \tab R (>= 2.8.1), flowCore\cr
  License: \tab Artistic-2.0\cr
  
 }
}
\author{
Mike Jiang \email{wjiang2@fhcrc.org},
Greg Finak \email{gfinak@fhcrc.org}

Maintainer: Mike Jiang \email{wjiang2@fhcrc.org}
}
\keyword{package}
