% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{ncdfFlowSet,flowSet-method}
\alias{ncdfFlowSet,flowSet-method}
\title{create ncdfFlowSet from flowSet}
\usage{
\S4method{ncdfFlowSet}{flowSet}(x, ncdfFile, dim = 2)
}
\arguments{
\item{x}{\code{flowSet}}

\item{ncdfFile}{\code{character} specifies the file name of cdf file}

\item{dim}{\code{integer} see details in \link{read.ncdfFlowset}.}
}
\description{
Normally the \code{ncdfFlowSet} is constructed by loading raw FCS files using \code{read.ncdfFlowSet}.
In case there is a legacy \code{flowSet} object, we can convert it to \code{ncdfFlowSet} with this constructor.
}
\examples{
data(GvHD)
fs <- GvHD[1:2]
ncfs <- ncdfFlowSet(fs)
}

