% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Subset,ncdfFlowSet,filterResultList-method}
\alias{Subset,ncdfFlowList,filter-method}
\alias{Subset,ncdfFlowList,filterResultList-method}
\alias{Subset,ncdfFlowSet,filter-method}
\alias{Subset,ncdfFlowSet,filterResultList-method}
\alias{Subset,ncdfFlowSet,list-method}
\title{subset a ncdfFlowSet by filter}
\usage{
\S4method{Subset}{ncdfFlowSet,filterResultList}(x, subset, select, ...)

\S4method{Subset}{ncdfFlowList,filterResultList}(x, subset, select, ...)

\S4method{Subset}{ncdfFlowSet,filter}(x, subset, ...)

\S4method{Subset}{ncdfFlowList,filter}(x, subset, ...)

\S4method{Subset}{ncdfFlowSet,list}(x, subset, select, validityCheck = TRUE,
  ...)
}
\arguments{
\item{x}{\code{ncdfFlowSet} or \code{ncdfFlowList}}

\item{subset,select,...}{see \code{\link[flowCore]{Subset-methods}}}

\item{validityCheck}{\code{logical} whether to skip validity check for speed.}
}
\value{
one or more \code{ncdfFlowSet} objects which share the same hdf5 file with the original one.
}
\description{
Equivalent to \code{Subset} method for \code{flowSet}.
}

