% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{rbind2,ncdfFlowSet,ncdfFlowSet-method}
\alias{rbind2,ncdfFlowList,ANY-method}
\alias{rbind2,ncdfFlowSet,flowFrame-method}
\alias{rbind2,ncdfFlowSet,ncdfFlowSet-method}
\title{combine multiple ncdfFlowSet objects into one}
\usage{
\S4method{rbind2}{ncdfFlowSet,ncdfFlowSet}(x, y, ncdfFile = tempfile(pattern =
  "ncfs"))

\S4method{rbind2}{ncdfFlowSet,flowFrame}(x, y)

\S4method{rbind2}{ncdfFlowList,ANY}(x, ncdfFile = tempfile(pattern = "ncfs"))
}
\arguments{
\item{x}{any \R object, typically matrix-like.}

\item{y}{any \R object, typically similar to \code{x}, or missing
    completely.}
}
\value{
a new ncdfFlowSet with a new cdf file that combines multiple raw datasets.
}
\description{
Similar to \code{\link[=rbind2,flowSet,flowSet-method]{flowCore:rbind2}}.
But one needs to first construct a ncdfFlowList and then apply rbind2 to it instead of merging them pairwise
}
\examples{
data(GvHD)

nc1 <- ncdfFlowSet(GvHD[1:2])
nc2 <- ncdfFlowSet(GvHD[3:4])
nc3 <- ncdfFlowSet(GvHD[5:6])
ncfslist <- as(list(nc1,nc2,nc3),"ncdfFlowList")
nc4 <- rbind2(ncfslist)
nc4
}

