% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{ncdfFlowSet,flowFrame-method}
\alias{ncdfFlowSet,flowFrame-method}
\alias{ncdfFlowSet,flowSet-method}
\alias{read.ncdfFlowSet}
\title{create ncdfFlowSet from flowFrame (not supported)}
\usage{
\S4method{ncdfFlowSet}{flowFrame}(x, ncdfFile)

\S4method{ncdfFlowSet}{flowSet}(x, ncdfFile)

read.ncdfFlowSet(files = NULL, ncdfFile, flowSetId = "",
  isWriteSlice = TRUE, phenoData, channels = NULL, ...)
}
\arguments{
\item{x}{\code{flowSet}}

\item{ncdfFile}{\code{character} specifies the file name of cdf file}

\item{files}{A character vector giving the source FCS raw file paths.}

\item{ncdfFile}{A character scalar giving the output file name. Default is NULL and the function will generate a random
file in the temporary folder, potentially adding the \code{.cdf} suffix unless a file
extension is already present. It is sometimes useful to specify this file path to avoid the failure of writing large flow data set to cdf file
due to the the shortage of disk space in system temporary folder.
It is only valid when isNewNcFile=TRUE}

\item{flowSetId}{A character scalar giving the unique ncdfFlowSet ID.}

\item{isWriteSlice}{A logical scalar indicating whether the raw data should also be copied.if FALSE,
an empty cdf file is created with the dimensions (sample*events*channels) supplied by raw FCS files.}

\item{phenoData}{An object of \code{AnnotatedDataFrame} providing a way to manually set the phenotyoic data for the whole data set in ncdfFlowSet.}

\item{channels}{A character vector specifying which channels to extract from FCS files.
It can be useful when FCS files do not share exactly the same channel names.
Thus this argument is used to select those common channels that are of interests.
Default value is NULL and the function will try to scan the FCS headers of all files
and determine the common channels.}

\item{...}{extra arguments to be passed to \code{\link{read.FCS}}.}
}
\value{
A ncdfFlowSet object
}
\description{
create ncdfFlowSet from flowFrame (not supported)

Normally the \code{ncdfFlowSet} is constructed by loading raw FCS files using \code{read.ncdfFlowSet}.
In case there is a legacy \code{flowSet} object, we can convert it to \code{ncdfFlowSet} with this constructor.

read FCS files from the disk and load them into a ncdfFlowSet object
}
\examples{
data(GvHD)
fs <- GvHD[1:2]
ncfs <- ncdfFlowSet(fs)
library(ncdfFlow)

path<-system.file("extdata","compdata","data",package="flowCore")
files<-list.files(path,full.names=TRUE)[1:3]

#create ncdfFlowSet from fcs with the actual raw data written in cdf
nc1  <- read.ncdfFlowSet(files=files,ncdfFile="ncfsTest.nc",flowSetId="fs1",isWriteSlice= TRUE)
nc1
nc1[[1]]
unlink(nc1)
rm(nc1)

#create empty ncdfFlowSet from fcs and add data slices afterwards
nc1  <- read.ncdfFlowSet(files=files,ncdfFile="ncfsTest.nc",flowSetId="fs1",isWriteSlice= FALSE)
fs1<-read.flowSet(files)
nc1[[1]] <- fs1[[1]]
nc1[[1]]
nc1[[2]]
}
\seealso{
\code{\link{clone.ncdfFlowSet}}
}

