% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{[[,ncdfFlowSet,ANY-method}
\alias{[[,ncdfFlowSet,ANY-method}
\title{extract a \code{flowFrame} object from \code{ncdfFlowSet}}
\usage{
\S4method{[[}{ncdfFlowSet,ANY}(x, i, j, use.exprs = TRUE, ...)
}
\arguments{
\item{x}{a \code{ncdfFlowSet}}

\item{i}{a \code{numeric} or \code{character} used as sample index}

\item{j}{a \code{numeric} or \code{character} used as channel index}

\item{use.exprs}{a \code{logical} scalar indicating whether to read the actual data from cdf}

\item{...}{other arguments. not used.}
}
\description{
Simliar to \code{\link[=[[,flowSet-method]{[[}}, and there are cerntain ways to
reduce the disk IO and optimize the speed.
}
\examples{
data(GvHD)
nc <- ncdfFlowSet(GvHD[1:2])
samples <- sampleNames(nc)
sn <- samples[1]
#return the entire flowFrame
fr <- nc[[sn]]

#access the flowFrame meta data without loading the raw event data from disk
nc[[sn, use.exprs = FALSE]]

#only read a subset of channels (more efficient than reading entire data set)
nc[[sn, 1:2]]
}

