% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-functions.R
\name{subset.ncdfFlowSet}
\alias{subset.ncdfFlowList}
\alias{subset.ncdfFlowSet}
\title{subset the ncdfFlowSet/ncdfFlowList based on 'pData'}
\usage{
\method{subset}{ncdfFlowSet}(x, subset, ...)

\method{subset}{ncdfFlowList}(x, subset, ...)
}
\arguments{
\item{x}{\code{ncdfFlowSet} or \code{ncdfFlowList}}

\item{subset}{logical expression(within the context of pData) indicating samples to keep. see \code{\link[base:subset]{subset}}}

\item{...}{other arguments. (not used)}
}
\value{
a subset of code{ncdfFlowSet} or \code{ncdfFlowList} object
}
\description{
subset the ncdfFlowSet/ncdfFlowList based on 'pData'
}

