% Generated by roxygen2 (4.0.1): do not edit by hand
\name{subset.ncdfFlowList}
\alias{subset.ncdfFlowList}
\alias{subset.ncdfFlowSet}
\title{subset the ncdfFlowSet/ncdfFlowList based on 'pData'}
\usage{
\method{subset}{ncdfFlowList}(x, ...)

\method{subset}{ncdfFlowSet}(x, subset)
}
\arguments{
\item{...}{other arguments}

\item{x}{\code{ncdfFlowSet} or \code{ncdfFlowList}}

\item{subset}{logical expression(within the context of pData) indicating samples to keep. see \link{base:subset}}
}
\value{
a subset of code{ncdfFlowSet} or \code{ncdfFlowList} object
}
\description{
subset the ncdfFlowSet/ncdfFlowList based on 'pData'
}

