% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{unlink,ncdfFlowSet-method}
\alias{unlink,ncdfFlowSet-method}
\title{delete the cdf file associated with the ncdfFlowSet object}
\usage{
\S4method{unlink}{ncdfFlowSet}(x, recursive = FALSE, force = FALSE)
}
\arguments{
\item{x}{\code{ncdfFlowSet}}

\item{recursive}{see \link[base:unlink]{unlink}}

\item{force}{see \link[base:unlink]{unlink}}
}
\description{
ncdfFlowSet object is unrecoverable after cdf is deleted.
So this method is usually called when ncdfFlowSet object is no longer in need.
}
\examples{
data(GvHD)
nc <- ncdfFlowSet(GvHD[1:2])
nc[[1]] # data is loaded from cdf file
unlink(nc)
}

