% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{updateIndices,ncdfFlowSet,character,logical-method}
\alias{updateIndices,ncdfFlowSet,character,logical-method}
\title{update the event indices of the target sample in ncdfFlowSet}
\usage{
\S4method{updateIndices}{ncdfFlowSet,character,logical}(x, y, z)
}
\arguments{
\item{z}{a \code{logical} vector}
}
\description{
update the event indices of the target sample in ncdfFlowSet
}

