% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{initIndices,ncdfFlowSet-method}
\alias{initIndices}
\alias{initIndices,ncdfFlowSet-method}
\title{initialize the event indices for the entire ncdfFlowSet with NA}
\usage{
\S4method{initIndices}{ncdfFlowSet}(x)
}
\arguments{
\item{x}{\code{ncdfFlowSet} object}
}
\description{
For internal use.
}

