% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfFlowList-methods.R
\name{lapply,ncdfFlowList-method}
\alias{lapply,ncdfFlowList-method}
\title{lapply method for ncdfFlowList}
\usage{
\S4method{lapply}{ncdfFlowList}(X, FUN, level = 2, ...)
}
\arguments{
\item{X}{\code{ncdfFlowList} object}

\item{FUN}{\code{function} to apply}

\item{level}{\code{numeric}. It controls whether loop at `ncdfFlowSet` level or `sample` level. 
when level = 2 (default value),\code{FUN} is applied to each sample. When level = 1, \code{FUN} is applied to each object stored in \code{data} slot.}

\item{...}{other arguments passed to \code{FUN}}
}
\description{
Depending on \code{level} parameter, loop either iterates through the list of ncdfFlowSet objects 
or every\code{flowFrame} objects.
}
