% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{ncdfFlowList-class}
\alias{[,ncdfFlowList,ANY-method}
\alias{[[,ncdfFlowList,character-method}
\alias{[[,ncdfFlowList,logical-method}
\alias{[[,ncdfFlowList,numeric-method}
\alias{colnames,ncdfFlowList-method}
\alias{densityplot,formula,ncdfFlowList-method}
\alias{filter,ncdfFlowList,filter-method}
\alias{length,ncdfFlowList-method}
\alias{ncdfFlowList-class}
\alias{pData,ncdfFlowList-method}
\alias{pData<-,ncdfFlowList,data.frame-method}
\alias{phenoData,ncdfFlowList-method}
\alias{phenoData<-,ncdfFlowList,AnnotatedDataFrame-method}
\alias{sampleNames,ncdfFlowList-method}
\alias{show,ncdfFlowList-method}
\alias{split,ncdfFlowList,character-method}
\alias{split,ncdfFlowList,factor-method}
\alias{xyplot,formula,ncdfFlowList-method}
\title{a class that stores multiple ncdfFlowSet objects}
\usage{
\S4method{filter}{ncdfFlowList,filter}(x, filter, method = "missing",
  sides = "missing", circular = "missing", init = "missing")

\S4method{[[}{ncdfFlowList,numeric}(x, i, j, ...)

\S4method{[[}{ncdfFlowList,logical}(x, i, j, ...)

\S4method{[[}{ncdfFlowList,character}(x, i, j, ...)

\S4method{length}{ncdfFlowList}(x)

\S4method{show}{ncdfFlowList}(object)

\S4method{sampleNames}{ncdfFlowList}(object)

\S4method{[}{ncdfFlowList,ANY}(x, i, j, ..., drop = TRUE)

\S4method{split}{ncdfFlowList,factor}(x, f, drop = FALSE, ...)

\S4method{split}{ncdfFlowList,character}(x, f, drop = FALSE, ...)

\S4method{phenoData}{ncdfFlowList}(object)

\S4method{pData}{ncdfFlowList}(object)

\S4method{colnames}{ncdfFlowList}(x)

\S4method{xyplot}{formula,ncdfFlowList}(x, data, ...)

\S4method{densityplot}{formula,ncdfFlowList}(x, data, ...)
}
\arguments{
\item{x}{\code{ncdfFlowList} object}

\item{filter}{\code{filter} to be applied}

\item{method}{\code{missing} not used}

\item{sides}{\code{missing} not used}

\item{circular}{\code{missing} not used}

\item{init}{\code{missing} not used}

\item{i}{\code{numeric} index}

\item{j}{column index}

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link{drop}} for further details.
  }

\item{object}{Any R object}

\item{...}{further potential arguments passed to methods.}

\item{f}{a \sQuote{factor} in the sense that \code{\link{as.factor}(f)}
    defines the grouping, or a list of such factors in which case their
    interaction is used for the grouping.}

\item{data}{
    For the \code{formula} methods, a data frame (or more precisely,
    anything that is a valid \code{envir} argument in
    \code{\link{eval}}, e.g., a list or an environment) containing values
    for any variables in the formula, as well as \code{groups} and
    \code{subset} if applicable.  If not found in \code{data}, or if
    \code{data} is unspecified, the variables are looked for in the
    environment of the formula.  For other methods (where \code{x} is
    not a formula), \code{data} is usually ignored, often with a warning
    if it is explicitly specified.
  }
}
\description{
It is a list of ncdfFlowSet objects
}
\section{Objects from the Class}{

Objects can be created by coercing a list of ncdfFlowSet objects
as("ncdfFlowList",nclist = .... #a list of ncdfFlowSet objects)
}
\examples{
data(GvHD)
nc1 <- ncdfFlowSet(GvHD[1])
nc2 <- ncdfFlowSet(GvHD[2])
nc3 <- ncdfFlowSet(GvHD[3])
list1 <- list(nc1, nc2, nc3)
#coerce from list to ncdfFlowList
nclist <- as(list1, "ncdfFlowList")
nclist
#coerce(collapse) from ncdfFlowList to a single flowFrame
collapsedData <- as(nclist, "flowFrame")
collapsedData
}
\seealso{
\code{\link{ncdfFlowSet}}
}

