% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfFlowList-methods.R,
%   R/ncdfFlowSet-split-methods.R
\docType{methods}
\name{split,ncdfFlowList,factor-method}
\alias{split,ncdfFlowList,factor-method}
\alias{split,ncdfFlowList,character-method}
\alias{split,ncdfFlowSet,filter-method}
\alias{split,ncdfFlowSet,filterResultList-method}
\alias{split,ncdfFlowSet,list-method}
\alias{split,ncdfFlowSet,factor-method}
\alias{split,ncdfFlowSet,character-method}
\title{split a \code{ncdfFlowSet} object.}
\usage{
\S4method{split}{ncdfFlowList,factor}(x, f, drop = FALSE, ...)

\S4method{split}{ncdfFlowList,character}(x, f, drop = FALSE, ...)

\S4method{split}{ncdfFlowSet,filter}(x, f, drop = FALSE, population = NULL,
  prefix = NULL, ...)

\S4method{split}{ncdfFlowSet,filterResultList}(x, f, drop = FALSE,
  population = NULL, prefix = NULL, ...)

\S4method{split}{ncdfFlowSet,list}(x, f, isNew = FALSE, drop = FALSE,
  population = NULL, prefix = NULL, ...)

\S4method{split}{ncdfFlowSet,factor}(x, f, isNew = FALSE, drop = FALSE, ...)

\S4method{split}{ncdfFlowSet,character}(x, f, drop = FALSE, ...)
}
\arguments{
\item{x}{\code{ncdfFlowSet}}

\item{f, drop, population, prefix, ...}{see \code{\link{split-methods}}}

\item{isNew}{\code{logical} wehther to create a new hdf file or using existing hdf file.}
}
\value{
a list of \code{ncdfFlowSet} objects that may not may not share the same hdf file depending on \code{isNew} argument.
}
\description{
Equivalent to \code{split} method for \code{flowSet} object.
}
