% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ncdfFlowSet-accessors.R
\docType{methods}
\name{ncfsApply,ncdfFlowSet-method}
\alias{ncfsApply}
\alias{ncfsApply,ncdfFlowSet-method}
\title{apply method for ncdfFlowSet (for internal use)}
\usage{
\S4method{ncfsApply}{ncdfFlowSet}(x, FUN, ..., use.exprs = FALSE,
  newNcFile = NULL)
}
\arguments{
\item{x}{\code{ncdfFlowSet}}

\item{FUN}{\code{function} to apply}

\item{...}{other arguments to pass to \code{FUN}}

\item{use.exprs}{\code{logical} see \code{\link{fsApply}}}

\item{newNcFile}{\code{logical} wether to create a new hdf file or simply overwrite the existing file.}
}
\description{
It is equivalent to \code{\link{fsApply}}. But the latter could cause memory issue
when \code{FUN} returns a \code{flowFrame}. \code{ncdfApply} writes to a new cdf file instead of memory.
Thus it will return a ncdfFlowSet object.
}
\details{
When the function given by argument "FUN" does not return the entire flowFrame object with the same
size of the original one (such as compensate,transform...), \code{\link[flowCore:fsApply]{fsApply}} should be used instead.
}
\examples{
data(GvHD)
nc <- ncdfFlowSet(GvHD[1:2])

#use fsApply when FUN does not return a flowFrame
fsApply(nc, nrow)
fsApply(nc, range)

#use ncfsApply when FUN returns a flowFrame
lgcl <- logicleTransform( w = 0.5, t= 10000, m =4.5)
nc1 <- ncfsApply(nc, transform, `FL1-H` = lgcl(`FL1-H`), `FL2-H` = lgcl(`FL2-H`))
}

