% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{densityplot,formula,ncdfFlowSet-method}
\alias{densityplot,formula,ncdfFlowList-method}
\alias{densityplot,formula,ncdfFlowSet-method}
\alias{xyplot,formula,ncdfFlowList-method}
\alias{xyplot,formula,ncdfFlowSet-method}
\title{flowViz plot methods.}
\usage{
\S4method{densityplot}{formula,ncdfFlowSet}(x, data, ...)

\S4method{densityplot}{formula,ncdfFlowList}(x, data, ...)

\S4method{xyplot}{formula,ncdfFlowSet}(x, data, ...)

\S4method{xyplot}{formula,ncdfFlowList}(x, data, ...)
}
\arguments{
\item{x}{\code{formula}}

\item{data}{\code{ncdfFlowSet} or \code{ncdfFlowList}}

\item{...}{other arguments passed to \code{flowViz}}
}
\description{
flowViz plot methods.
}

