% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{[,ncdfFlowSet,ANY-method}
\alias{[,ncdfFlowList,ANY-method}
\alias{[,ncdfFlowSet,ANY-method}
\title{subsetting by sampleNames,channels(not for events) methods}
\usage{
\S4method{[}{ncdfFlowSet,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{ncdfFlowList,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{\code{ncdfFlowSet}}

\item{i}{sample index(or name)}

\item{j}{column(or channel) index (or name)}

\item{...}{other arguments not used}

\item{drop}{\code{logical} not used.}
}
\description{
similar to \code{\link[=[,flowSet-method]{[}}.
}
\examples{
data(GvHD)
nc <- ncdfFlowSet(GvHD[1:2])
samples <- sampleNames(nc)
nc[1]
nc1 <- nc[samples[1]]
#nc1 and nc share the cdf file
all.equal(getFileName(nc1), getFileName(nc))
}

