% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{updateIndices,ncdfFlowSet,character,logical-method}
\alias{updateIndices}
\alias{updateIndices,ncdfFlowSet,character,logical-method}
\title{update the event indices of the target sample in ncdfFlowSet}
\usage{
\S4method{updateIndices}{ncdfFlowSet,character,logical}(x, y, z)
}
\arguments{
\item{x}{\code{ncdfFlowSet} object}

\item{y}{\code{character} sample name}

\item{z}{\code{logical} vector to be assigned.}
}
\description{
For internal use.
}

